Move app-routing.module.ts to its own directory and use routes import to consolidate code
